For now, send TCP segments over the connection, not bytes.
I haven't implemented the OSI protocols underneath TCP yet, so not only is it a gross misuse of TCP for it to be in-charge of sending raw bytes over the connection, I am making things unnecessarily harder on myself to serialize / deserialize the segments.  I want to just be focusing on the underlying RFC 9293 spec for now.

Even though this was a stopping point for me for a while, it was actually useful as I now have a better understanding of the boundaries of the protocols!