fix: use sync.Once to stop Connection close race condition
Before this commit, it would've been possible for a gateway and host to try to close a connection at the same time.  It would've resulted in a panic as it would try and close a closed channel.

Now, when a connection is closed, any modules using it should end their receive channel loop, ensure both channels of the multiplexed connection have been closed, and set the physical interface's connection, send, and receive to nil, ready to be reinitialized if necessary.