fix: control loop for synchronisation of module actions
Prior to this commit, the only real synchronisation was on the Connection object during a close operation.  sync.Once would ensure a Connection's channel close would only occur once on that channel.  This was useful if e.g. a host and gateway both closed and both tried to end the same Connection object.

However, operations like sending would not be able to tell if the channel was closed in a thread-safe way.  Also, listening on a physical interface was done in a Module method which didn't have any mutexes.  Listening is now done in the physical interface.

The two options to implement synchronisation were either use some sort of mutual exclusion zone in the code or have a `select` control loop.  Preferring to "share memory by communicating" rather than "communicating by sharing memory", the solution uses the 'done' channel pattern, which closes a 'done' channel on the Connection rather than the send/receive channels themselves; `select` statements are used instead of mutexes here.

This also simplifies physical interface connection management, as a physical interface's connection doesn't have to be recreated with a new reconnect call, which would require a mutex.  Instead, a new connection can be sent over a channel and it will replace the either currently dead connection, or the currently active connection.

A possible disadvantage with this implementation is that if the done channel has been closed external to either the sender or receiver, but the sender and receiver are actively communicating, the `done` channel in their selects may not be pseudo-randomly chosen for a while, so even though the connection has technically been specified as broken, they could continue communication.

This could be remediated by not using the 'done' channel pattern and instead forcefully setting channels to `nil` and handling any panics.