fix: gouroutine safety
Prior to this commit, there was no real 'thread' safety among goroutines for operations on e.g. a module's physical interface map or when changing connections between two modules.

Multiplexed connections now share not only the same `done` channel, but also a shared 'destroy' sync.Once, as before they didn't, which made it possible for the `done` channel to be closed by the first connection and closed again by the second connection close.

Physical interfaces can mimic the real world of unsafely destroying a connection instantly, without needing to rely on Go's pseudo-random goroutine selection in a `select` statement to hope the `done` channel closure broadcast will be selected before trying to read off a channel that `done` says should actually be closed.

Physical interfaces also start with a dummy / sentinel connection which simplifies logic so no branching is necessary to check the validility of a connection reference.