fix: remove deadlock on unconnected send
Previously, a newly created physical interface would have a nil send channel which would mean that during a send operation on an unconnected module, it would deadlock.  This is because a new physical interface is created with a nil send and receive channel in the connection and an unbuffered `done` channel; unbuffered to simplify mutex logic with a sentinel Connection.

This commit initialises a new physical interface with an unbuffered `done` channel, a nil receive channel, and a closed send channel.