feat: bind addresses to interfaces and continue listen if connection changes
Previously, if a passive listen took place but then the connection on an interface was changed, it would be necessary to e.g. have to respecify each module should passive listen on that physical interface again.

Now, if the module was already passively listening on that interface and a connection change occurs, the physical interface will continue to passively listen.