fix: possible deadlock on interface connection close
If multiple routines try to e.g. set a connection (which indirectly closes the active connection) or directly close a connection at the same time, many routines could be held waiting for the wg channel to become available, by which time a new connection may have already been established and when it enters the 'guarded' code, it will be waiting indefinitely for the new connection to shutdown which it may never do.